+++ fuente/composicion/composicion.ts
@@ -1,4 +1,4 @@
-//Junta los cuerpos, interacciones, entorno, casos límite y renderizado.
+//Junta los cuerpos, interacciones, entorno, casos límite y Dibujante.
 //Debería estar acá la creación de canvas y contexto??
 
 import { Cuadricula } from "../cuadricula/cuadricula";
@@ -8,13 +8,14 @@ import { Forma } from "../geometria-plana/formas";
 import { Entorno } from "../interaccion/entorno";
 import { QuadTree } from "../interaccion/quad-tree";
 import { Restriccion } from "../interaccion/restriccion";
-import { Renderizado } from "../renderizado/renderizado";
+import { Dibujante } from "../renderizado/dibujante";
+import { Punto } from "../tipos/tipos";
 import { Tiempo } from "./tiempo";
 
 export class Composicion {
 
-    /**Herramienta renderizadora.*/
-    render: Renderizado;
+    /**Herramienta Dibujante.*/
+    dibujante: Dibujante;
     /**Conjunto de cuerpos sobre los que trabaja la composición.*/
     cuerpos: Cuerpo[] = []
     /**Conjunto de formas sobre las que trabaja la composición.*/
@@ -33,11 +34,45 @@ export class Composicion {
 
     private constructor(canvas?: HTMLCanvasElement, idCanvas?: string) {
         if (canvas) {
-            this.render = Renderizado.crearConCanvas(canvas);
+            this.dibujante = Dibujante.crearConCanvas(canvas);
         }
         else {
-            this.render = Renderizado.crearConIdCanvas(idCanvas!);
+            this.dibujante = Dibujante.crearConIdCanvas(idCanvas!);
         }
+    }
+    /**Retorna la medida horizontal del canvas.*/
+    get anchoCanvas(): number {
+        return this.dibujante.anchoCanvas;
+    }
+
+    /**Retorna la media vertical del canvas. */
+    get altoCanvas(): number {
+        return this.dibujante.altoCanvas;
+    }
+
+    /**Retorna un punto ubicado en el centro del canvas.*/
+    get centroCanvas(): Punto {
+        return { x: this.dibujante.anchoCanvas / 2, y: this.dibujante.altoCanvas / 2 };
+    }
+
+    /**Retorna el color del canvas.*/
+    get colorCanvas(): string | undefined {
+        return this.dibujante.colorCanvas
+    }
+
+    /**Modifica la medida horizontal del canvas.*/
+    set anchoCanvas(ancho: number) {
+        this.dibujante.anchoCanvas = ancho;
+    }
+
+    /**Modifica la medida vertical del canvas. */
+    set altoCanvas(alto: number) {
+        this.dibujante.altoCanvas = alto;
+    }
+
+    /**Modifica el color del canvas.*/
+    set colorCanvas(color: string) {
+        this.dibujante.colorCanvas = color;
     }
 
     set entorno(entorno: Entorno) {
@@ -49,7 +84,7 @@ export class Composicion {
     }
 
     /**Retorna un objeto de tipo Composicion a partir del id de un canvas.*/
-    static crearConIDCanvas(idCanvas: string): Composicion {
+    static crearConIdCanvas(idCanvas: string): Composicion {
         const nuevaCompo: Composicion = new Composicion(undefined, idCanvas)
         return nuevaCompo;
     }
@@ -63,8 +98,8 @@ export class Composicion {
 
     /**Define el ancho y el alto del canvas, en pixeles. */
     tamanoCanvas(ancho: number, alto: number) {
-        this.render.anchoCanvas = ancho;
-        this.render.altoCanvas = alto;
+        this.dibujante.anchoCanvas = ancho;
+        this.dibujante.altoCanvas = alto;
     }
 
     /**Agrega cuerpos al conjunto de cuerpos manipulados por la composición. */
@@ -82,13 +117,13 @@ export class Composicion {
         // Interaccion.reboteEntreCuerpos(this.cuerpos)
         let niveles: number = this.nivelesQuadTree;
         let capacidad: number = Math.ceil(this.cuerpos.length / (2 ** niveles))
-        const Quad: QuadTree = new QuadTree(0, 0, this.render.anchoCanvas, this.render.altoCanvas, capacidad, niveles);
+        const Quad: QuadTree = new QuadTree(0, 0, this.dibujante.anchoCanvas, this.dibujante.altoCanvas, capacidad, niveles);
         for (let cuerpo of this.cuerpos) {
             Quad.insertarPunto(cuerpo.posicion, cuerpo)
         }
         Quad.reboteEslasticoCuerpos()
         if (this.trazarQuadTree) {
-            Quad.trazar(this.render, { colorTrazo: 'skyblue' })
+            Quad.trazar(this.dibujante, { colorTrazo: 'skyblue' })
         }
     }
 
@@ -97,51 +132,51 @@ export class Composicion {
         // Interaccion.contactoSimple(this.cuerpos)
         let niveles: number = 9;
         let capacidad: number = Math.ceil(this.cuerpos.length / (2 ** niveles))
-        const Quad: QuadTree = new QuadTree(0, 0, this.render.anchoCanvas, this.render.altoCanvas, capacidad, niveles);
+        const Quad: QuadTree = new QuadTree(0, 0, this.dibujante.anchoCanvas, this.dibujante.altoCanvas, capacidad, niveles);
         for (let cuerpo of this.cuerpos) {
             Quad.insertarPunto(cuerpo.posicion, cuerpo)
         }
         Quad.contactoSimpleCuerpos()
         if (this.trazarQuadTree) {
-            Quad.trazar(this.render, { colorTrazo: 'skybule' })
+            Quad.trazar(this.dibujante, { colorTrazo: 'skybule' })
         }
     }
 
     /**Método gráfico. Pinta el interior de los cuerpos de la composición en el canvas.*/
     rellenarCuerpos() {
-        this.render.rellenarFormas(this.cuerpos)
+        this.dibujante.rellenarFormas(this.cuerpos)
     }
 
     /**Método gráfico. Traza los cuerpos de la composición en el canvas.*/
     trazarCuerpos() {
-        this.render.trazarFormas(this.cuerpos)
+        this.dibujante.trazarFormas(this.cuerpos)
     }
 
     /**Método gráfico. Pinta y/o rellena los cuerpos de la composición, según lo definido para cada cuerpo.*/
-    renderizarCuerpos() {
-        this.render.renderizarFormas(this.cuerpos)
+    dibujarCuerpos() {
+        this.dibujante.dibujarFormas(this.cuerpos)
     }
 
     /**Método gráfico. Pinta y/o rellena las formas de la composición, según lo definido para cada forma.*/
-    renderizarFormas() {
-        this.render.renderizarFormas(this.formas)
+    dibujarFormas() {
+        this.dibujante.dibujarFormas(this.formas)
     }
 
     /**Crea un loop para ejecutar dos funciones, una asociada a la duración de cada tick y otra a los fps.          
      * El atributo .tick permite cambiar su duración en milisegundos.       
      * La propiedad .fps permite ajustar su número.         
      */
-    animacion(funcionCalcular: () => void, funcionRenderizar: () => void): void {
+    animacion(funcionCalcular: () => void, funciondibujar: () => void): void {
         let tiempoCalculo: Tiempo = new Tiempo()
         let tiempoFrame: Tiempo = new Tiempo()
 
         const funcionAnimar = () => {
             if (this.animar && !this.usarfpsNativos) {
                 tiempoCalculo.iterarPorSegundo(funcionCalcular, 1000 / this.tick)
-                tiempoFrame.iterarPorSegundo(funcionRenderizar, this.fps)
+                tiempoFrame.iterarPorSegundo(funciondibujar, this.fps)
             } else if (this.animar && this.usarfpsNativos) {
                 tiempoCalculo.iterarPorSegundo(funcionCalcular, 1000 / this.tick)
-                funcionRenderizar();
+                funciondibujar();
             }
             requestAnimationFrame(funcionAnimar)
         }
